Write the end-to-end import that osm2streets will use. Don't refactor too much yet
